Cleaned CLI functionality. Added offline pyshark install call.
