+++ Wireshark/FilterDataModel.py
@@ -1,0 +1,3 @@
+class packageFiltered:
+	def packagefiltered():
+		print "Constructor called"
